test: adapt workflow behavior tests for Cloud
